ai-service/models/model_manager.py
"""
Model Manager for VidyutAI AI Service.
Handles loading, inference, and management of ML models.
"""

import os
import logging
import joblib
import numpy as np
from typing import Dict, List, Any, Optional, Union
import pandas as pd
from datetime import datetime

from config.settings import settings

logger = logging.getLogger(__name__)

class ModelManager:
    """
    Manages machine learning models for the AI service.
    """
    
    def __init__(self):
        """Initialize the model manager."""
        logger.info("Initializing ModelManager")
        self.models = {}
        self.load_models()
    
    def load_models(self):
        """Load all required models."""
        try:
            # In a real implementation, this would load saved models
            # For now, we'll just log the process
            logger.info("Loading anomaly detection model")
            # self.models['anomaly_detection'] = joblib.load(
            #     os.path.join(settings.MODEL_PATH, settings.ANOMALY_DETECTION_MODEL)
            # )
            
            logger.info("Loading prediction model")
            # self.models['prediction'] = joblib.load(
            #     os.path.join(settings.MODEL_PATH, settings.PREDICTION_MODEL)
            # )
            
            # For demonstration, we'll create placeholder models
            from sklearn.ensemble import IsolationForest
            self.models['anomaly_detection'] = IsolationForest(
                n_estimators=100, 
                contamination=0.02, 
                random_state=42
            )
            
            # Placeholder for prediction model
            self.models['prediction'] = None
            
            logger.info("Models loaded successfully")
        
        except Exception as e:
            logger.error(f"Error loading models: {str(e)}")
            # Continue with empty models dictionary
            # In production, you might want to raise an exception or implement retry logic
    
    def detect_anomalies(self, features: pd.DataFrame) -> Dict[str, np.ndarray]:
        """
        Detect anomalies in the provided features.
        
        Args:
            features: DataFrame with extracted features
            
        Returns:
            Dictionary with anomaly scores and flags
        """
        try:
            if 'anomaly_detection' not in self.models:
                logger.warning("Anomaly detection model not loaded")
                # Return placeholder results
                return {
                    'anomaly_score': np.zeros(len(features)),
                    'is_anomaly': np.zeros(len(features), dtype=bool)
                }
            
            # In a real implementation, this would use the loaded model
            # For now, generate random results
            anomaly_scores = np.random.normal(-0.5, 0.2, size=len(features))
            is_anomaly = anomaly_scores < -0.8
            
            return {
                'anomaly_score': anomaly_scores,
                'is_anomaly': is_anomaly
            }
        
        except Exception as e:
            logger.error(f"Error in anomaly detection: {str(e)}")
            # Return empty results
            return {
                'anomaly_score': np.zeros(len(features)),
                'is_anomaly': np.zeros(len(features), dtype=bool)
            }
    
    def predict(self, features: pd.DataFrame) -> np.ndarray:
        """
        Make predictions using the loaded prediction model.
        
        Args:
            features: DataFrame with extracted features
            
        Returns:
            Array of predictions
        """
        try:
            if 'prediction' not in self.models:
                logger.warning("Prediction model not loaded")
                # Return placeholder predictions
                return np.ones(len(features)) * 100.0
            
            # In a real implementation, this would use the loaded model
            # For now, generate synthetic predictions
            base_value = 100.0
            hour_effect = features['hour'].values * 2.0
            day_effect = features['day_of_week'].values * 5.0
            
            predictions = base_value + hour_effect - day_effect + np.random.normal(0, 5, size=len(features))
            
            return predictions
        
        except Exception as e:
            logger.error(f"Error in prediction: {str(e)}")
            # Return placeholder predictions
            return np.ones(len(features)) * 100.0
    
    def get_model_info(self) -> Dict[str, Any]:
        """
        Get information about the loaded models.
        
        Returns:
            Dictionary with model information
        """
        try:
            return {
                'anomaly_detection': {
                    'type': 'Isolation Forest' if 'anomaly_detection' in self.models else 'Not loaded',
                    'loaded': 'anomaly_detection' in self.models,
                    'last_updated': datetime.now().isoformat()
                },
                'prediction': {
                    'type': 'LSTM' if 'prediction' in self.models else 'Not loaded',
                    'loaded': 'prediction' in self.models,
                    'last_updated': datetime.now().isoformat()
                }
            }
        
        except Exception as e:
            logger.error(f"Error getting model info: {str(e)}")
            return {
                'error': str(e)
            }
    
    def retrain_model(self, model_type: str, training_data: pd.DataFrame) -> bool:
        """
        Retrain a specific model with new data.
        
        Args:
            model_type: Type of model to retrain ('anomaly_detection' or 'prediction')
            training_data: DataFrame with training data
            
        Returns:
            True if retraining was successful, False otherwise
        """
        try:
            logger.info(f"Retraining {model_type} model")
            
            # In a real implementation, this would retrain the model
            # For now, just log the process
            logger.info(f"Retraining {model_type} model with {len(training_data)} samples")
            
            # Placeholder for retraining logic
            if model_type == 'anomaly_detection':
                # Retrain anomaly detection model
                pass
            elif model_type == 'prediction':
                # Retrain prediction model
                pass
            else:
                logger.warning(f"Unknown model type: {model_type}")
                return False
            
            logger.info(f"{model_type} model retrained successfully")
            return True
        
        except Exception as e:
            logger.error(f"Error retraining {model_type} model: {str(e)}")
            return False