ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
langchain==0.1.1
langchain-openai==0.0.2
openai==1.7.2
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2
pandas==2.0.3
python-dotenv==1.0.0
mangum==0.17.0
